Add DifficultNounsPanel, StatisticsPanel.
